Rendering is now done by Hazel::Scene
No more custom rendering code.  Its all done by the Hazel Engine Scene object.
This does mean the it's limited to colored squares (as that is all the scene renderer does at the moment), so you will need to use your imagination!
The scene camera is controlled via a C++ "script" (i.e. a NativeScriptComponent on the camera entity).